Updated MCU Schematic
The original schematic had a major error in it, where SW1 was shorted to GND on button press. 

Long story short, before I added the power-on-reset cap there was a resistor connecting reset to ground. When I added the cap, I removed the resistor (but in editing the schematic, I didn't catch that the connection to GND remained).